moved map to where the display prompt is
