+++ index.html
@@ -160,17 +160,7 @@
                     <div class="control-btn b">B</div>
                 </div>
             </div>
-            <div id="right-panel" class="right">
-                <canvas
-                    id="minimap"
-                    style="
-                        position: absolute;
-                        top: 10px;
-                        right: 10px;
-                        display: block;
-                    "
-                ></canvas>
-            </div>
+            <div id="right-panel" class="right"></div>
             <div id="footer" class="bottom">
                 <div class="footer-head-info">
                     <div class="control-btns d-desktop-hide">
@@ -220,6 +210,7 @@
                             <div id="energy-bar-remain"></div>
                         </div>
                     </div>
+                    <canvas id="minimap" style="display: block"></canvas>
                 </div>
             </div>
         </div>
